fix(next): Fix issue where annotated tags lead to a version bump, no matter which commit keyword is used
The issue was caused, because the hash of a tag was used as a starting point for finding newer commits. The hash of a lightweight tag is the associated commits hash. The hash of an annotated tag does not belong to any commit. Now, the hash of an associated commit is used to find newer commits.